+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/DefaultComponents.kt
@@ -24,6 +24,7 @@ import mozilla.components.browser.menu.item.BrowserMenuItemToolbar
 import mozilla.components.browser.menu.item.SimpleBrowserMenuItem
 import mozilla.components.browser.session.storage.SessionStorage
 import mozilla.components.browser.state.engine.EngineMiddleware
+import mozilla.components.browser.state.engine.middleware.SessionPrioritizationMiddleware
 import mozilla.components.browser.state.selector.selectedTab
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.browser.storage.sync.PlacesHistoryStorage
@@ -162,7 +163,8 @@ open class DefaultComponents(private val applicationContext: Context) {
                 SearchMiddleware(applicationContext),
                 RecordingDevicesMiddleware(applicationContext),
                 LastAccessMiddleware(),
-                PromptMiddleware()
+                PromptMiddleware(),
+                SessionPrioritizationMiddleware()
             ) + EngineMiddleware.create(engine)
         )
     }